+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [BREAKING] Merged `concurrent` feature with `std` (#974).
 * [BREAKING] Changed `TransactionRequest` to use expected output recipients instead of output notes (#976).
 * [BREAKING] Removed `TransactionExecutor` from `Client` and `NoteScreener` (#998).
+* [BREAKING] Renamed P2IDR to P2IDE (#1016).
 * Replaced deprecated #[clap(...)] with #[command(...)] and #[arg(...)] (#897).
 
 ### Features